requirements.txt
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.1.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.136.1
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.4.0
gitdb==4.0.12
GitPython==3.1.50
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mpmath==1.3.0
narwhals==2.21.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.2
pandas==3.0.2
pillow==12.2.0
polars==1.40.1
polars-runtime-32==1.40.1
protobuf==7.34.1
psutil==7.2.2
pyarrow==24.0.0
pydantic==2.13.4
pydantic_core==2.46.4
pydeck==0.9.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.27
python-telegram-bot==22.7
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
smmap==5.0.3
starlette==1.0.0
streamlit==1.57.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
ultralytics==8.4.47
ultralytics-thop==2.0.19
unicorn==2.1.4
urllib3==2.7.0
uvicorn==0.46.0
watchdog==6.0.0
websockets==16.0
